+++ src/core/alien/kernels/simple_csr/CSRStructInfo.h
@@ -95,7 +95,10 @@ class CSRStructInfo
     return *this;
   }
 
-  CSRStructInfo* clone() const { return new CSRStructInfo(*this); }
+  CSRStructInfo* clone() const
+  {
+    return new CSRStructInfo(*this);
+  }
 
   void init(Arccore::Integer nrow)
   {
@@ -115,58 +118,115 @@ class CSRStructInfo
     m_cols.resize(nnz);
   }
 
-  Arccore::Int64 timestamp() const { return m_timestamp; }
+  Arccore::Int64 timestamp() const
+  {
+    return m_timestamp;
+  }
 
-  void setTimestamp(Arccore::Int64 value) { m_timestamp = value; }
+  void setTimestamp(Arccore::Int64 value)
+  {
+    m_timestamp = value;
+  }
 
-  bool getSymmetric() const { return m_symmetric; }
+  bool getSymmetric() const
+  {
+    return m_symmetric;
+  }
 
-  void setSymmetric(bool value) { m_symmetric = value; }
+  void setSymmetric(bool value)
+  {
+    m_symmetric = value;
+  }
 
-  Integer getNRows() const { return m_nrow; }
+  Integer getNRows() const
+  {
+    return m_nrow;
+  }
 
-  Integer getNRow() const { return m_nrow; }
+  Integer getNRow() const
+  {
+    return m_nrow;
+  }
 
   Arccore::UniqueArray<Arccore::Integer>& getBlockRowOffset()
   {
     return m_block_row_offset;
   }
 
-  Arccore::ArrayView<Integer> getRowOffset() { return m_row_offset.view(); }
+  Arccore::ArrayView<Integer> getRowOffset()
+  {
+    return m_row_offset.view();
+  }
 
   Arccore::ConstArrayView<Integer> getRowOffset() const
   {
     return m_row_offset.constView();
   }
 
-  const int* kcol() const { return m_row_offset.data(); }
+  const int* kcol() const
+  {
+    return m_row_offset.data();
+  }
 
-  int* kcol() { return m_row_offset.data(); }
+  int* kcol()
+  {
+    return m_row_offset.data();
+  }
 
   ConstArrayView<Integer> getBlockRowOffset() const
   {
     return m_block_row_offset.constView();
   }
 
-  UniqueArray<Integer>& getCols() { return m_cols; }
+  UniqueArray<Integer>& getCols()
+  {
+    return m_cols;
+  }
 
-  UniqueArray<Integer>& getBlockCols() { return m_block_cols; }
+  UniqueArray<Integer>& getBlockCols()
+  {
+    return m_block_cols;
+  }
 
-  ConstArrayView<Integer> getCols() const { return m_cols.constView(); }
+  ConstArrayView<Integer> getCols() const
+  {
+    return m_cols.constView();
+  }
 
-  int* cols() { return m_cols.data(); }
+  int* cols()
+  {
+    return m_cols.data();
+  }
 
-  const int* cols() const { return m_cols.data(); }
+  const int* cols() const
+  {
+    return m_cols.data();
+  }
 
-  ConstArrayView<Integer> getBlockCols() const { return m_block_cols.constView(); }
+  ConstArrayView<Integer> getBlockCols() const
+  {
+    return m_block_cols.constView();
+  }
 
-  ColOrdering& getColOrdering() { return m_col_ordering; }
+  ColOrdering& getColOrdering()
+  {
+    return m_col_ordering;
+  }
 
-  ColOrdering getColOrdering() const { return m_col_ordering; }
+  ColOrdering getColOrdering() const
+  {
+    return m_col_ordering;
+  }
 
-  void setDiagFirst(bool val) { m_diag_first = val; }
+  void setDiagFirst(bool val)
+  {
+    m_diag_first = val;
+  }
 
-  bool getDiagFirstOpt() const { return m_diag_first; }
+  bool getDiagFirstOpt() const
+  {
+    return m_diag_first;
+  }
 
   Integer getRowSize(Integer row) const
   {
@@ -178,11 +238,20 @@ class CSRStructInfo
     return m_block_row_offset[row + 1] - m_block_row_offset[row];
   }
 
-  Integer getNnz() const { return m_row_offset[m_nrow]; }
+  Integer getNnz() const
+  {
+    return m_row_offset[m_nrow];
+  }
 
-  Integer getNElems() const { return m_row_offset[m_nrow]; }
+  Integer getNElems() const
+  {
+    return m_row_offset[m_nrow];
+  }
 
-  Integer getBlockNnz() const { return m_block_row_offset[m_nrow]; }
+  Integer getBlockNnz() const
+  {
+    return m_block_row_offset[m_nrow];
+  }
 
   UniqueArray<Integer>& getUpperDiagOffset()
   {
@@ -210,9 +279,10 @@ class CSRStructInfo
 
   void allocate()
   {
-    m_cols.resize(m_row_offset[m_nrow]);
+    auto row_offset = (m_nrow > 0) ? m_row_offset[m_nrow] : 0;
+    m_cols.resize(row_offset);
     if (m_is_variable_block)
-      m_block_cols.resize(m_row_offset[m_nrow]);
+      m_block_cols.resize(row_offset);
   }
 
   void computeUpperDiagOffset() const